requirements.txt
numpy
pandas
scipy
pyTelegramBotAPI
requests
beautifulsoup4
python-dotenv
psycopg2-binary